Finished editing the right panels
The right panel UI is completed for both panels and are also both
connected to cards.
Found and fixed a bug that didn't remove the card from the Player's hand
once the drag item was destroyed.
Did some resizing pivoting so that dimensions for UI panels change
correctly when resized.
Bug: the health bars on the left panel aren't added correctly...