Started implementation of Enemy player
Added the Enemy player - will randomly put down a minion card in its
hand when it's not the Player's turn. (bug: for some reason, its deck
doesn't run out of cards...)
Added idNumber to PlayingCard to differentiate different cards with the
same name and stats.
Added lines for the grid.
Added the Knight card (need a prefab for this).
Added a GetCardType function to Base/PlayerCard.
Added a GetCardOfType function to Player and Enemy so they can randomly
select a card in the hand that is of that type.