Finished implementing skills
Finished writing the code for all the skills; they are now fully
functional.
Use: Drag the skill onto the target.
For Lure, you place on the "area" of the lure (must contain a card, if
the card is a player, then surrounding enemy cards will lose health each
turn). It's kind of like, you're giving the target card the lure and
they distribute it to the surrounding cards.
Notes:
Not yet allowed the Enemy to use skills.
Skills are still not side-specific (can use heal on the Enemy, can
attack Player cards).
Skills will NOT work if placed on an empty square - they still appear as
a card on the board.